+++ messaging-core/src/test/java/com/nabla/sdk/messaging/core/data/stubs/NablaMessagingClientStub.kt
@@ -47,10 +47,10 @@ class NablaMessagingClientStub(
 
     override suspend fun createConversation(
         title: String?,
-        providerIdToAssign: Uuid?,
+        providerIds: List<Uuid>?,
     ): Result<Conversation> {
         return runCatchingCancellable {
-            conversationRepository.createConversation(title, providerIdToAssign)
+            conversationRepository.createConversation(title, providerIds)
         }
     }